backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
litellm==1.34.0
requests==2.31.0
pyyaml==6.0.1
google-cloud-aiplatform==1.126.0

# Vector store dependencies (sentence-transformers + FAISS)
faiss-cpu==1.7.4
numpy==1.24.3
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.28.1

